apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
boto3==1.34.157
redis==5.0.8
rq==1.16.2
stripe==10.9.0
python-multipart==0.0.9
aiofiles==24.1.0
anthropic==0.34.2
openai==1.46.0
google-generativeai==0.7.2
assemblyai==0.30.0
resend==2.3.0
sentry-sdk[fastapi]==2.14.0
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.2
